refactor: centralize bootstrap helpers
